Refactor enterprise WeChatClient
