+++ wechatpy/enterprise/client/api/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+from __future__ import absolute_import, unicode_literals
+
+from .department import WeChatDepartment  # NOQA
+from .media import WeChatMedia  # NOQA
+from .message import WeChatMessage  # NOQA
+from .menu import WeChatMenu  # NOQA
+from .tag import WeChatTag  # NOQA
+from .user import WeChatUser  # NOQA
